refactor(iot): delete api/iot/common hub, narrow to internal/shared (#575 B7a)
api/iot/common coupled every IoT sub-package to a grab-bag hub. Replaced:
- Multi-domain ErrorRenderer + RecordUnregisteredTagScan + the two
  PyrePortal error-message constants move to api/iot/internal/shared —
  Go's internal-package fence limits it to api/iot/..., and it no longer
  re-exports generic error constructors.
- Capacity + StudentAlreadyActive conflict types/renderers move to
  api/iot/checkin (the only flow that emits them), with their tests and
  the B0 wire goldens.
- The Error*/RenderError pass-through shims are deleted; handlers call
  api/common directly. NormalizeTagID shim deleted; callers use
  users.NormalizeTagID.

Wire bytes proven unchanged by the wire goldens, PyrePortal guard, and
the route/auth-matrix goldens.
