+++ backend/api/iot/checkin/capacity_errors.go
@@ -1,4 +1,17 @@
-package common
+package checkin
+
+// Capacity and duplicate-visit conflict responses for the device checkin
+// flow. Moved verbatim from the deleted api/iot/common in issue #575 B7 —
+// checkin is the only flow that emits them.
+//
+// WIRE CONTRACT (PyrePortal): the Code values, Message strings, and details
+// field names below are substring-matched by the kiosk
+// (PyrePortal/src/services/apiErrors.ts) and pinned byte-for-byte by
+// wire_format_test.go. Note the activity-capacity details keys are
+// current_occupancy/max_capacity while PyrePortal reads
+// current_participants/max_participants — a known mismatch preserved
+// deliberately; fixing it requires a coordinated backend+PyrePortal change
+// (tracked separately, issue #575 follow-up).
 
 import (
 	"errors"
@@ -7,30 +20,13 @@ import (
 	"time"
 
 	"github.com/go-chi/render"
-	"github.com/moto-nrw/project-phoenix/api/common"
 	activeSvc "github.com/moto-nrw/project-phoenix/services/active"
-	feedbackSvc "github.com/moto-nrw/project-phoenix/services/feedback"
-	iotSvc "github.com/moto-nrw/project-phoenix/services/iot"
 )
-
-// RenderError renders an error response and logs any render failures.
-// For server errors (5xx), it delegates to common.RenderError which handles
-// slog logging and Sentry capture. For non-5xx errors, it renders directly.
-// Exported for use by sub-packages (devices, checkin, etc.)
-func RenderError(w http.ResponseWriter, r *http.Request, renderer render.Renderer) {
-	common.RenderError(w, r, renderer)
-}
 
 // Common error variables
 var (
 	ErrRoomCapacityExceeded     = errors.New("room capacity exceeded")
 	ErrActivityCapacityExceeded = errors.New("activity capacity exceeded")
-)
-
-// Error message constants for reuse across handlers
-const (
-	ErrMsgPersonNotStudent = "person is not a student"
-	ErrMsgRFIDTagNotFound  = "RFID tag not found"
 )
 
 // RoomCapacityExceededError represents detailed information about a capacity exceeded error
@@ -182,166 +178,5 @@ func ErrorStudentAlreadyActive(studentID, existingVisitID int64, entryTime *time
 			RoomID:          roomID,
 			RoomName:        roomName,
 		},
-	}
-}
-
-// ErrorInvalidRequest returns a 400 Bad Request error response
-func ErrorInvalidRequest(err error) render.Renderer {
-	return common.ErrorInvalidRequest(err)
-}
-
-// ErrorInternalServer returns a 500 Internal Server Error response
-func ErrorInternalServer(err error) render.Renderer {
-	return common.ErrorInternalServer(err)
-}
-
-// ErrorInternalServerWrap returns a 500 response with a stable client-facing message
-// while preserving the full error chain for Sentry and slog.
-func ErrorInternalServerWrap(clientMsg string, cause error) render.Renderer {
-	return common.ErrorInternalServerWrap(clientMsg, cause)
-}
-
-// ErrorNotFound returns a 404 Not Found error response
-func ErrorNotFound(err error) render.Renderer {
-	return common.ErrorNotFound(err)
-}
-
-// ErrorConflict returns a 409 Conflict error response
-func ErrorConflict(err error) render.Renderer {
-	return common.ErrorConflict(err)
-}
-
-// ErrorRenderer renders an error to an HTTP response based on the IoT service error type
-func ErrorRenderer(err error) render.Renderer {
-	// Delegate to service-specific error handlers
-	if iotErr, ok := err.(*iotSvc.IoTError); ok {
-		return handleIoTServiceError(iotErr)
-	}
-
-	if activeErr, ok := err.(*activeSvc.ActiveError); ok {
-		return handleActiveServiceError(activeErr)
-	}
-
-	if feedbackErr, ok := err.(*feedbackSvc.InvalidEntryDataError); ok {
-		return ErrorInvalidRequest(feedbackErr)
-	}
-
-	if renderer := handleFeedbackServiceError(err); renderer != nil {
-		return renderer
-	}
-
-	// For unknown errors, return a generic internal server error
-	return ErrorInternalServer(err)
-}
-
-// handleIoTServiceError maps IoT service errors to HTTP responses
-func handleIoTServiceError(iotErr *iotSvc.IoTError) render.Renderer {
-	switch iotErr.Unwrap() {
-	case iotSvc.ErrDeviceNotFound:
-		return ErrorNotFound(iotErr)
-	case iotSvc.ErrInvalidDeviceData:
-		return ErrorInvalidRequest(iotErr)
-	case iotSvc.ErrDuplicateDeviceID:
-		return ErrorConflict(iotErr)
-	case iotSvc.ErrInvalidStatus:
-		return ErrorInvalidRequest(iotErr)
-	case iotSvc.ErrDeviceProtected:
-		return common.ErrorForbidden(iotErr)
-	case iotSvc.ErrDeviceOffline:
-		return ErrorConflict(iotErr)
-	case iotSvc.ErrNetworkScanFailed:
-		return ErrorInternalServer(iotErr)
-	case iotSvc.ErrDatabaseOperation:
-		return ErrorInternalServer(iotErr)
-	default:
-		return handleIoTErrorTypes(iotErr)
-	}
-}
-
-// handleIoTErrorTypes handles specific IoT error types
-func handleIoTErrorTypes(iotErr *iotSvc.IoTError) render.Renderer {
-	switch iotErr.Err.(type) {
-	case *iotSvc.DeviceNotFoundError:
-		return ErrorNotFound(iotErr)
-	case *iotSvc.DuplicateDeviceIDError:
-		return ErrorConflict(iotErr)
-	default:
-		return ErrorInternalServer(iotErr)
-	}
-}
-
-// handleActiveServiceError maps Active service errors to HTTP responses
-func handleActiveServiceError(activeErr *activeSvc.ActiveError) render.Renderer {
-	// Check for conflict errors (409)
-	if isActiveConflictError(activeErr.Err) {
-		return ErrorConflict(activeErr)
-	}
-
-	// Check for not found errors (404)
-	if isActiveNotFoundError(activeErr.Err) {
-		return ErrorNotFound(activeErr)
-	}
-
-	// Check for validation errors (400)
-	if isActiveValidationError(activeErr.Err) {
-		return ErrorInvalidRequest(activeErr)
-	}
-
-	// Check for database errors (500)
-	if errors.Is(activeErr.Err, activeSvc.ErrDatabaseOperation) {
-		return ErrorInternalServer(activeErr)
-	}
-
-	// Default to internal server error
-	return ErrorInternalServer(activeErr)
-}
-
-// isActiveConflictError checks if error is a conflict error
-func isActiveConflictError(err error) bool {
-	return errors.Is(err, activeSvc.ErrRoomConflict) ||
-		errors.Is(err, activeSvc.ErrSessionConflict) ||
-		errors.Is(err, activeSvc.ErrStudentAlreadyInGroup) ||
-		errors.Is(err, activeSvc.ErrGroupAlreadyInCombination) ||
-		errors.Is(err, activeSvc.ErrStudentAlreadyActive) ||
-		errors.Is(err, activeSvc.ErrStaffAlreadySupervising) ||
-		errors.Is(err, activeSvc.ErrDeviceAlreadyActive)
-}
-
-// isActiveNotFoundError checks if error is a not found error
-func isActiveNotFoundError(err error) bool {
-	return errors.Is(err, activeSvc.ErrActiveGroupNotFound) ||
-		errors.Is(err, activeSvc.ErrVisitNotFound) ||
-		errors.Is(err, activeSvc.ErrGroupSupervisorNotFound) ||
-		errors.Is(err, activeSvc.ErrCombinedGroupNotFound) ||
-		errors.Is(err, activeSvc.ErrGroupMappingNotFound) ||
-		errors.Is(err, activeSvc.ErrNoActiveSession) ||
-		errors.Is(err, activeSvc.ErrStaffNotFound)
-}
-
-// isActiveValidationError checks if error is a validation error
-func isActiveValidationError(err error) bool {
-	return errors.Is(err, activeSvc.ErrActiveGroupAlreadyEnded) ||
-		errors.Is(err, activeSvc.ErrVisitAlreadyEnded) ||
-		errors.Is(err, activeSvc.ErrSupervisionAlreadyEnded) ||
-		errors.Is(err, activeSvc.ErrCombinedGroupAlreadyEnded) ||
-		errors.Is(err, activeSvc.ErrInvalidTimeRange) ||
-		errors.Is(err, activeSvc.ErrCannotDeleteActiveGroup) ||
-		errors.Is(err, activeSvc.ErrInvalidData) ||
-		errors.Is(err, activeSvc.ErrInvalidActivitySession)
-}
-
-// handleFeedbackServiceError maps Feedback service errors to HTTP responses
-func handleFeedbackServiceError(err error) render.Renderer {
-	switch {
-	case errors.Is(err, feedbackSvc.ErrEntryNotFound):
-		return ErrorNotFound(err)
-	case errors.Is(err, feedbackSvc.ErrInvalidEntryData):
-		return ErrorInvalidRequest(err)
-	case errors.Is(err, feedbackSvc.ErrStudentNotFound):
-		return ErrorNotFound(err)
-	case errors.Is(err, feedbackSvc.ErrInvalidDateRange):
-		return ErrorInvalidRequest(err)
-	default:
-		return nil // Not a feedback error
 	}
 }
