+++ backend/api/iot/internal/shared/unregistered_tags.go
@@ -1,4 +1,4 @@
-package common
+package shared
 
 import (
 	"context"
