+++ backend/api/iot/sessions/timeout.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/go-chi/render"
 	"github.com/moto-nrw/project-phoenix/api/common"
-	iotCommon "github.com/moto-nrw/project-phoenix/api/iot/common"
+	shared "github.com/moto-nrw/project-phoenix/api/iot/internal/shared"
 	"github.com/moto-nrw/project-phoenix/auth/device"
 )
 
@@ -18,7 +18,7 @@ func (rs *Resource) processSessionTimeout(w http.ResponseWriter, r *http.Request
 	// Process timeout via device ID
 	result, err := rs.ActiveService.ProcessSessionTimeout(r.Context(), deviceCtx.ID)
 	if err != nil {
-		iotCommon.RenderError(w, r, iotCommon.ErrorRenderer(err))
+		common.RenderError(w, r, shared.ErrorRenderer(err))
 		return
 	}
 
@@ -40,20 +40,20 @@ func (rs *Resource) updateSessionActivity(w http.ResponseWriter, r *http.Request
 
 	var req SessionActivityRequest
 	if err := render.Bind(r, &req); err != nil {
-		iotCommon.RenderError(w, r, iotCommon.ErrorRenderer(err))
+		common.RenderError(w, r, shared.ErrorRenderer(err))
 		return
 	}
 
 	// Get current session for this device
 	session, err := rs.ActiveService.GetDeviceCurrentSession(r.Context(), deviceCtx.ID)
 	if err != nil {
-		iotCommon.RenderError(w, r, iotCommon.ErrorRenderer(err))
+		common.RenderError(w, r, shared.ErrorRenderer(err))
 		return
 	}
 
 	// Update session activity
 	if err := rs.ActiveService.UpdateSessionActivity(r.Context(), session.ID); err != nil {
-		iotCommon.RenderError(w, r, iotCommon.ErrorRenderer(err))
+		common.RenderError(w, r, shared.ErrorRenderer(err))
 		return
 	}
 
@@ -73,13 +73,13 @@ func (rs *Resource) validateSessionTimeout(w http.ResponseWriter, r *http.Reques
 
 	var req TimeoutValidationRequest
 	if err := render.Bind(r, &req); err != nil {
-		iotCommon.RenderError(w, r, iotCommon.ErrorRenderer(err))
+		common.RenderError(w, r, shared.ErrorRenderer(err))
 		return
 	}
 
 	// Validate the timeout request
 	if err := rs.ActiveService.ValidateSessionTimeout(r.Context(), deviceCtx.ID, req.TimeoutMinutes); err != nil {
-		iotCommon.RenderError(w, r, iotCommon.ErrorRenderer(err))
+		common.RenderError(w, r, shared.ErrorRenderer(err))
 		return
 	}
 
@@ -99,7 +99,7 @@ func (rs *Resource) getSessionTimeoutInfo(w http.ResponseWriter, r *http.Request
 
 	info, err := rs.ActiveService.GetSessionTimeoutInfo(r.Context(), deviceCtx.ID)
 	if err != nil {
-		iotCommon.RenderError(w, r, iotCommon.ErrorRenderer(err))
+		common.RenderError(w, r, shared.ErrorRenderer(err))
 		return
 	}
